backend/requirements.txt
fastapi
uvicorn
groq
python-docx
docxtpl